// api/src/__generated__/users.ts
/**
 * !!! This file is autogenerated do not edit by hand !!!
 *
 * Generated by: @databases/pg-schema-print-types
 * Checksum: UEjwWfXx9G/1FBXNHML01XpvmkdlBhFiYmCkvig+JQeUCJlkEOOIc72BIYZe+yJPqbUN4ZeWJczbX4DJIZde1A==
 */

// eslint:disable
// tslint:disable

interface Users {
  api_token: (string) | null
  /**
   * @default CURRENT_TIMESTAMP
   */
  created_at: (Date) | null
  email: string
  email_verified_at: (Date) | null
  /**
   * @default nextval('users_id_seq'::regclass)
   */
  id: number & {readonly __brand?: 'users_id'}
  name: string
  password: string
  remember_token: (string) | null
  /**
   * @default CURRENT_TIMESTAMP
   */
  updated_at: (Date) | null
}
export default Users;

interface Users_InsertParameters {
  api_token?: (string) | null
  /**
   * @default CURRENT_TIMESTAMP
   */
  created_at?: (Date) | null
  email: string
  email_verified_at?: (Date) | null
  /**
   * @default nextval('users_id_seq'::regclass)
   */
  id?: number & {readonly __brand?: 'users_id'}
  name: string
  password: string
  remember_token?: (string) | null
  /**
   * @default CURRENT_TIMESTAMP
   */
  updated_at?: (Date) | null
}
export type {Users_InsertParameters}